requirements.txt
python-telegram-bot==21.10
requests==2.31.0
matplotlib==3.8.0
Pillow==10.1.0
PyGithub==2.3.0
python-dotenv==1.0.0